+++ lib/Imagine/Filter/Basic/Save.php
@@ -35,7 +35,7 @@ class Save implements FilterInterface
      * @param string $path
      * @param array  $options
      */
-    public function __construct($path, array $options = array())
+    public function __construct($path = null, array $options = array())
     {
         $this->path    = $path;
         $this->options = $options;