+++ lib/Imagine/Filter/Basic/Thumbnail.php
@@ -31,15 +31,22 @@ class Thumbnail implements FilterInterface
     private $mode;
 
     /**
+     * @var string
+     */
+    private $filter;
+
+    /**
      * Constructs the Thumbnail filter with given width, height and mode
      *
      * @param BoxInterface $size
      * @param string       $mode
+     * @param string       $filter
      */
-    public function __construct(BoxInterface $size, $mode = ImageInterface::THUMBNAIL_INSET)
+    public function __construct(BoxInterface $size, $mode = ImageInterface::THUMBNAIL_INSET, $filter = ImageInterface::FILTER_UNDEFINED)
     {
         $this->size = $size;
         $this->mode = $mode;
+        $this->filter = $filter;
     }
 
     /**
@@ -47,6 +54,6 @@ class Thumbnail implements FilterInterface
      */
     public function apply(ImageInterface $image)
     {
-        return $image->thumbnail($this->size, $this->mode);
+        return $image->thumbnail($this->size, $this->mode, $this->filter);
     }
 }