+++ lib/Imagine/Filter/Transformation.php
@@ -44,6 +44,11 @@ final class Transformation implements FilterInterface, ManipulatorInterface
     private $filters = array();
 
     /**
+     * @var array
+     */
+    private $sorted;
+
+    /**
      * An ImagineInterface instance.
      *
      * @var ImagineInterface
@@ -86,12 +91,27 @@ final class Transformation implements FilterInterface, ManipulatorInterface
     }
 
     /**
+     * Returns a list of filters sorted by their priority. Filters with same priority will be returned in the order they were added.
+     *
+     * @return array
+     */
+    public function getFilters()
+    {
+        if (null === $this->sorted) {
+            ksort($this->filters);
+            $this->sorted = call_user_func_array('array_merge', $this->filters);
+        }
+
+        return $this->sorted;
+    }
+
+    /**
      * {@inheritdoc}
      */
     public function apply(ImageInterface $image)
     {
         return array_reduce(
-            $this->filters,
+            $this->getFilters(),
             array($this, 'applyFilter'),
             $image
         );
@@ -180,7 +200,7 @@ final class Transformation implements FilterInterface, ManipulatorInterface
     /**
      * {@inheritdoc}
      */
-    public function save($path, array $options = array())
+    public function save($path = null, array $options = array())
     {
         return $this->add(new Save($path, $options));
     }
@@ -196,22 +216,23 @@ final class Transformation implements FilterInterface, ManipulatorInterface
     /**
      * {@inheritdoc}
      */
-    public function thumbnail(BoxInterface $size, $mode = ImageInterface::THUMBNAIL_INSET)
+    public function thumbnail(BoxInterface $size, $mode = ImageInterface::THUMBNAIL_INSET, $filter = ImageInterface::FILTER_UNDEFINED)
     {
-        return $this->add(new Thumbnail($size, $mode));
+        return $this->add(new Thumbnail($size, $mode, $filter));
     }
 
     /**
      * Registers a given FilterInterface in an internal array of filters for
      * later application to an instance of ImageInterface
      *
-     * @param FilterInterface $filter
-     *
+     * @param  FilterInterface $filter
+     * @param  int             $priority
      * @return Transformation
      */
-    public function add(FilterInterface $filter)
+    public function add(FilterInterface $filter, $priority = 0)
     {
-        $this->filters[] = $filter;
+        $this->filters[$priority][] = $filter;
+        $this->sorted = null;
 
         return $this;
     }