+++ lib/Imagine/Image/AbstractLayers.php
@@ -11,9 +11,6 @@
 
 namespace Imagine\Image;
 
-use Imagine\Image\LayersInterface;
-use Imagine\Image\ImageInterface;
-
 abstract class AbstractLayers implements LayersInterface
 {
     /**