+++ tests/Imagine/Test/Filter/TransformationTest.php
@@ -11,8 +11,10 @@
 
 namespace Imagine\Test\Filter;
 
+use Imagine\Filter\FilterInterface;
 use Imagine\Filter\Transformation;
 use Imagine\Image\Box;
+use Imagine\Image\ImageInterface;
 use Imagine\Image\Point;
 use Imagine\Image\ManipulatorInterface;
 
@@ -126,5 +128,49 @@ class TransformationTest extends FilterTestCase
         $transformation1->paste($transformation2->apply($img2), $start)
             ->show('png')
             ->apply($img1);
+    }
+
+    public function testFilterSorting()
+    {
+        $filter1 = new TestFilter();
+        $filter2 = new TestFilter();
+        $filter3 = new TestFilter();
+
+        $transformation1 = new Transformation();
+        $transformation1
+            ->add($filter1, 5)
+            ->add($filter2, -3)
+            ->add($filter3);
+
+        $expected1 = array(
+            $filter2,
+            $filter3,
+            $filter1,
+        );
+
+        $transformation2 = new Transformation();
+        $transformation2
+            ->add($filter1)
+            ->add($filter2)
+            ->add($filter3);
+
+        $expected2 = array(
+            $filter1,
+            $filter2,
+            $filter3,
+        );
+
+        $this->assertSame($expected1, $transformation1->getFilters());
+        $this->assertSame($expected2, $transformation2->getFilters());
+    }
+}
+
+class TestFilter implements FilterInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function apply(ImageInterface $image)
+    {
     }
 }